Cambios en el modulo de cafeteria

+++ database/migrations/2015_10_01_015100_create_MateriaPrima_table.php
@@ -15,8 +15,8 @@ class CreateMateriaPrimaTable extends Migration
         Schema::create('MateriaPrima', function(Blueprint $table) {
             $table->increments('id');
             $table->string('nombre');
-		  $table->integer('existencia');
-		  $table->float('precio');
+        	$table->integer('existencia');
+            $table->decimal('precio', 10, 2);
             $table->timestamps();
         });
     }
